--- test/test-applications/integrations/TestApplication.GraphQL.NativeSupport/Properties/launchSettings.json
@@ -1,20 +1,0 @@
-{
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:54567/",
-      "sslPort": 0
-    }
-  },
-  "profiles": {
-    "TestApplication.GraphQL.NativeSupport": {
-      "commandName": "Project",
-      "launchBrowser": false,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "http://localhost:54568/"
-    }
-  }
-}
