--- test/test-applications/integrations/TestApplication.GraphQL.NativeSupport/TestApplication.GraphQL.NativeSupport.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFrameworks>net7.0;net6.0</TargetFrameworks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="GraphQL" VersionOverride="$(LibraryVersion)" />
-    <PackageReference Include="GraphQL.MicrosoftDI" />
-    <PackageReference Include="GraphQL.Server.Transports.AspNetCore" />
-    <PackageReference Include="GraphQL.Server.Ui.Playground" />
-    <PackageReference Include="Microsoft.AspNetCore" />
-    <PackageReference Include="System.Diagnostics.DiagnosticSource" />
-    <PackageReference Include="System.Reactive" />
-  </ItemGroup>
-
-</Project>
