+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/LICENSE
