+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/StarWarsData.cs
