+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/StarWarsMutation.cs
