+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/StarWarsQuery.cs
