+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/StarWarsSchema.cs
