+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/StarWarsSubscription.cs
