+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/CharacterInterface.cs
