+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/DroidType.cs
