+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/EpisodeEnum.cs
