+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/HumanInputType.cs
