+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/HumanType.cs
