+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/StarWarsCharacter.cs
