+++ test/test-applications/integrations/TestApplication.GraphQL/TestApplication.GraphQL.csproj
@@ -1,20 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <LibraryVersion Condition="'$(LibraryVersion)' == ''" >2.4.0</LibraryVersion>
+    <TargetFrameworks>net7.0;net6.0</TargetFrameworks>
   </PropertyGroup>
-  
+
   <ItemGroup>
     <PackageReference Include="GraphQL" VersionOverride="$(LibraryVersion)" />
-    <PackageReference Include="GraphQL.StarWars" />
-    <PackageReference Include="GraphQL.Server.Transports.AspNetCore" VersionOverride="3.4.0" />
-    <PackageReference Include="GraphQL.Server.Ui.Playground" VersionOverride="3.4.0" />
+    <PackageReference Include="GraphQL.MicrosoftDI" />
+    <PackageReference Include="GraphQL.Server.Transports.AspNetCore" />
+    <PackageReference Include="GraphQL.Server.Ui.Playground" />
     <PackageReference Include="Microsoft.AspNetCore" />
-  </ItemGroup>
-
-  <!-- Reconfigure default usings template -->
-  <ItemGroup>
-    <Using Remove="System.Net.Http.Json"/>
+    <PackageReference Include="System.Diagnostics.DiagnosticSource" />
+    <PackageReference Include="System.Reactive" />
   </ItemGroup>
 
 </Project>
